升级播放器内核所需修改，顺带解决播放进度保存问题
