+++ app/src/main/java/com/github/tvbox/osc/player/IjkMediaPlayer.java
@@ -3,7 +3,7 @@ package com.github.tvbox.osc.player;
 import android.content.Context;
 import android.text.TextUtils;
 
-import com.dueeeke.videoplayer.ijk.IjkPlayer;
+import xyz.doikki.videoplayer.ijk.IjkPlayer;
 import com.github.tvbox.osc.api.ApiConfig;
 import com.github.tvbox.osc.bean.IJKCode;
