+++ app/src/main/java/com/github/tvbox/osc/player/controller/BoxVideoController.java
@@ -12,10 +12,10 @@ import androidx.annotation.AttrRes;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.dueeeke.videoplayer.controller.GestureVideoController;
-import com.dueeeke.videoplayer.controller.IControlComponent;
-import com.dueeeke.videoplayer.player.VideoView;
-import com.dueeeke.videoplayer.player.VideoViewManager;
+import xyz.doikki.videoplayer.controller.GestureVideoController;
+import xyz.doikki.videoplayer.controller.IControlComponent;
+import xyz.doikki.videoplayer.player.VideoView;
+import xyz.doikki.videoplayer.player.VideoViewManager;
 import com.github.tvbox.osc.R;
 
 import java.util.Map;
