+++ app/src/main/java/com/github/tvbox/osc/util/PlayerHelper.java
@@ -2,14 +2,14 @@ package com.github.tvbox.osc.util;
 
 import android.content.Context;
 
-import com.dueeeke.videoplayer.exo.ExoMediaPlayerFactory;
-import com.dueeeke.videoplayer.player.AndroidMediaPlayerFactory;
-import com.dueeeke.videoplayer.player.PlayerFactory;
-import com.dueeeke.videoplayer.player.VideoView;
-import com.dueeeke.videoplayer.player.VideoViewConfig;
-import com.dueeeke.videoplayer.player.VideoViewManager;
-import com.dueeeke.videoplayer.render.RenderViewFactory;
-import com.dueeeke.videoplayer.render.TextureRenderViewFactory;
+import xyz.doikki.videoplayer.exo.ExoMediaPlayerFactory;
+import xyz.doikki.videoplayer.player.AndroidMediaPlayerFactory;
+import xyz.doikki.videoplayer.player.PlayerFactory;
+import xyz.doikki.videoplayer.player.VideoView;
+import xyz.doikki.videoplayer.player.VideoViewConfig;
+import xyz.doikki.videoplayer.player.VideoViewManager;
+import xyz.doikki.videoplayer.render.RenderViewFactory;
+import xyz.doikki.videoplayer.render.TextureRenderViewFactory;
 import com.github.tvbox.osc.player.IjkMediaPlayer;
 import com.github.tvbox.osc.player.render.SurfaceRenderViewFactory;
 import com.orhanobut.hawk.Hawk;
@@ -108,7 +108,6 @@ public class PlayerHelper {
                 .setScreenScaleType(VideoView.SCREEN_SCALE_DEFAULT)
                 .setPlayerFactory(playerFactory)
                 .setRenderViewFactory(renderViewFactory)
-                .setProgressManager(new ProgressManagerImpl())
                 .build());
     }
 }
