--- app/src/main/res/layout/dkplayer_layout_vod_control_view.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <LinearLayout
-        android:id="@+id/bottom_container"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/dkplayer_controller_height"
-        android:layout_gravity="bottom"
-        android:gravity="center_vertical"
-        android:background="@drawable/dkplayer_shape_stardard_controller_bottom_bg"
-        android:orientation="horizontal">
-
-        <ImageView
-            android:id="@+id/iv_play"
-            android:layout_width="@dimen/dkplayer_controller_height"
-            android:layout_height="@dimen/dkplayer_controller_height"
-            android:padding="@dimen/dkplayer_controller_icon_padding"
-            android:src="@drawable/dkplayer_selector_play_button" />
-
-        <TextView
-            android:id="@+id/curr_time"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@android:color/white"
-            android:textSize="@dimen/dkplayer_controller_time_text_size"
-            tools:text="00:00" />
-
-        <SeekBar
-            android:id="@+id/seekBar"
-            android:layout_width="0dp"
-            android:layout_height="40dp"
-            android:layout_gravity="center_vertical"
-            android:layout_marginEnd="10dp"
-            android:layout_marginStart="10dp"
-            android:layout_weight="1"
-            android:background="@null"
-            android:max="1000"
-            android:maxHeight="2dp"
-            android:minHeight="2dp"
-            android:paddingEnd="0dp"
-            android:paddingStart="0dp"
-            android:progressDrawable="@drawable/dkplayer_layer_progress_bar"
-            android:thumb="@drawable/dkplayer_seekbar_thumb"
-            android:thumbOffset="0dp" />
-
-        <TextView
-            android:id="@+id/total_time"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@android:color/white"
-            android:textSize="@dimen/dkplayer_controller_time_text_size"
-            tools:text="00:00" />
-
-        <ImageView
-            android:id="@+id/fullscreen"
-            android:layout_width="@dimen/dkplayer_controller_height"
-            android:layout_height="@dimen/dkplayer_controller_height"
-            android:padding="@dimen/dkplayer_controller_icon_padding"
-            android:src="@drawable/dkplayer_selector_full_screen_button" />
-    </LinearLayout>
-
-    <ProgressBar
-        android:id="@+id/bottom_progress"
-        style="?android:attr/progressBarStyleHorizontal"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_gravity="bottom"
-        android:max="1000"
-        android:progressDrawable="@drawable/dkplayer_layer_progress_bar"
-        android:visibility="gone" />
-
-</FrameLayout>
